XWEB-001: Renaming files to avoid too repetitive word counter
