+++ src/features/counter/index.tsx
@@ -8,8 +8,8 @@ import {
   incrementAsync,
   incrementIfOdd,
   selectCount,
-} from './counterSlice'
-import styles from './Counter.module.css'
+} from './slice'
+import styles from './index.module.css'
 
 function Counter() {
   const dispatch = useAppDispatch()
