+++ src/features/counter/slice.ts
@@ -1,7 +1,7 @@
 import { createAsyncThunk, createSlice, PayloadAction } from '@reduxjs/toolkit'
 
 import type { AppState, AppThunk } from '#store'
-import { fetchCount } from './counterAPI'
+import { fetchCount } from './api'
 
 export interface CounterState {
   value: number
