+++ src/pages/index.tsx
@@ -1,7 +1,7 @@
 import type { NextPage } from 'next'
 import Head from 'next/head'
 
-import Counter from '#features/counter/Counter'
+import Counter from '#features/counter'
 import styles from '#styles/Home.module.css'
 
 const IndexPage: NextPage = () => {
