+++ src/store.ts
@@ -1,6 +1,6 @@
 import { configureStore, ThunkAction, Action } from '@reduxjs/toolkit'
 
-import counterReducer from '#features/counter/counterSlice'
+import counterReducer from '#features/counter/slice'
 
 export function makeStore() {
   return configureStore({
